+++ src/main/java/bot/commands/RoleAmongUs.java
@@ -6,7 +6,6 @@ import net.dv8tion.jda.api.entities.Member;
 import net.dv8tion.jda.api.entities.Message;
 import net.dv8tion.jda.api.entities.Role;
 import net.dv8tion.jda.api.entities.channel.unions.MessageChannelUnion;
-import net.dv8tion.jda.api.exceptions.ErrorResponseException;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -22,19 +21,11 @@ public class RoleAmongUs implements CommandExecutor {
         String content = message.getContentRaw();
         String[] args = content.split(" ");
         Guild guild = message.getGuild();
-        Member target;
+        Member target = args.length < 2 ? null : Bot.findMember(guild, args[1]);
         Role roleAmongUs = guild.getRoleById(Roles.ROLE_AMONG_US.id());
 
         if (member == null || !isMemberAllowed(member)) return;
         if (Channels.STAFF_AJUDANTES_CHANNEL.id() != channel.getIdLong()) return;
-
-        if (roleAmongUs == null) {
-            Bot.sendGhostMessage(channel,
-                    Messages.ERROR_REQUIRED_ROLES_NOT_FOUND.message(),
-                    10000);
-            message.delete().queue();
-            return;
-        }
 
         if (args.length < 2) {
             Bot.sendGhostMessage(channel, Messages.ERROR_TOO_FEW_ARGUMENTS.message(), 5000);
@@ -42,16 +33,26 @@ public class RoleAmongUs implements CommandExecutor {
             return;
         }
 
-        try {
-            target = guild.retrieveMemberById(args[1].replaceAll("[^0-9]+", "")).complete();
-        } catch (ErrorResponseException e) {
-            Bot.sendGhostMessage(channel, Messages.ERROR_MEMBER_NOT_FOUND.message(), 5000);
+        if (roleAmongUs == null) {
+            Bot.sendGhostMessage(channel, Messages.ERROR_REQUIRED_ROLES_NOT_FOUND.message(), 10000);
+            message.delete().queue();
+            return;
+        }
+
+        if (target == null) {
+            Bot.sendGhostMessage(channel, Messages.ERROR_MEMBER_NOT_FOUND.message(), 10000);
+            message.delete().queue();
+            return;
+        }
+
+        if (target.getUser().isBot()) {
+            Bot.sendGhostMessage(channel, "Um bot não pode receber este cargo.", 10000);
             message.delete().queue();
             return;
         }
 
         if (target.getRoles().contains(roleAmongUs)) {
-            channel.sendMessage("<@" + member.getIdLong() + "> o membro <@" + target.getIdLong() + "> já tem o cargo `Já Participou (Among Us)`.").queue();
+            channel.sendMessage("O membro <@" + target.getIdLong() + "> já tem o cargo `Já Participou (Among Us)`.").queue();
             message.delete().queue();
             return;
         }
@@ -61,12 +62,12 @@ public class RoleAmongUs implements CommandExecutor {
         channel.sendMessage("<@" + member.getIdLong() + "> o cargo `Já Participou (Among Us)` foi adicionado com sucesso à <@" + target.getIdLong() + ">!").queue();
     }
 
-    private static boolean isMemberAllowed(Member member) {
+    private boolean isMemberAllowed(Member member) {
         List<Role> possibleRoles = new ArrayList<>();
         Guild guild = member.getGuild();
         AtomicBoolean returned = new AtomicBoolean(false);
 
-        for (int i = 0; i < Roles.GENERAL_AJUDANTES.ids().size()-2; i++)
+        for (int i = 0; i < Roles.GENERAL_AJUDANTES.ids().size() - 2; i++)
             possibleRoles.add(guild.getRoleById(Roles.GENERAL_AJUDANTES.ids().get(i)));
 
         possibleRoles.forEach(r -> {
